New Seed MAP UPDATE
In the seed map, I fix the errors, where the land is the color blue for some reason, in addition adding code like edges for the land and adding rivers code. Now, the plan was to use the island as the mountains instead, and no water at all. However, I would like to talk about the river as the source code for the generative seed map use pathfinder ai algorithm, called A* to find the path of the rivers by avoiding collision, this code could be use for the AI for the enemy isntead on making rivers only