Tower can shoot
So I added a new script so it can shoot, first it is called Tower.cs plus projectile.cs. For now, when the tower shoots, the ant is dead instantly, later this will change for the better. sadly though I can really show how large the are of shooting as I realise the tower has it's own script and finding out how might take a lot of time actually, for now I want to ensure that the shooting works and just need a minor tweek of adding economy and damage, making it already in a playable state.